Move serdes helpers to data_transfer and pylint
