+++ api/python/t4/exceptions.py
@@ -3,4 +3,3 @@
 class PackageException(Exception):
     """ Exception relating to package validity. """
     pass
-
